Moved user examples into the core examples file.
